Show how much of a queue failed
A bare failure count says nothing without the number of jobs it is out of, so
the queue list now carries the share as well.

Reading it:
- the rate is the failed jobs against the total the list already shows
- green below one percent, orange below ten, red above that
- a queue that has never seen a job has no rate rather than a rate of zero

Sorting it:
- the column sorts like every other one, and queues without jobs stay last
- numbers are compared as floats now, or every rate below one percent would
  collapse into a single bucket and fall back to alphabetical order

+++ Dto/SortCriteria.php
@@ -213,7 +213,7 @@ abstract class SortCriteria
             }
 
             $result = $numeric
-                ? ((int) $leftValue <=> (int) $rightValue)
+                ? ((float) $leftValue <=> (float) $rightValue)
                 : strcmp((string) $leftValue, (string) $rightValue);
 
             // the column shows the opposite of what is compared here, so the
